Dark Souls Remastered: Restore Debug Camera (#83)
Greetings, I was tampering with the model viewer a while back and I managed to arbitrarily call a function that handled the debug camera making it accessible throughout the game. The only thing that really does is activating the camera which spawns at zero coordinates of the map allowing free movement while freezing both entity animation and special effects.