Mod Support Patch for Persona 3: Dancing in Moonlight, Persona 4: Dancing All Night, and Persona 5: Dancing in Starlight (#152)
This is a port of the mod support patches for these games from the
xdelta format to the format used here. This makes things a lot more
easier for me. I could only test the US versions for P3D and P5D and the
EU version for P4D so the patches are set for those versions
specifically.